Make GoalSubsystem subclass Subsystem.
[ci skip-rust]

[ci skip-build-wheels]
